+++ apps/gitness/src/main.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from 'react-dom/client'
 
-import App from './AppV1'
+import App from './AppV2'
 
 createRoot(document.getElementById('root')!).render(<App />)
